requirements.txt

fastapi==0.110.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.25.1
urllib3==1.26.18
uvicorn==0.29.0
websockets==12.0